reference for fields reworked + extra
Instead of `[::field.errors ::field.error]`, `[::field :errors :error]` is
used instead. This conforms with `[::field :help]` which would is a
lookup into the field in the form for the key `:help`. The main
advantage is that we now have a semi-uniform way of handling
extensions. `[::field :errors ...]` should be implemented as a generic
solution that would take everything and then have a function handle
it. This way ez-form would be fully extendible.

The `[::field <:keyword-here>]` has the added advantage that it
replaces data with data, so anything that then renders the hiccup, can
then use the new data to what's needed there.

`assert-fields` was also removed, as it places an unnecessary
constraint that doesn't give any superior benefit. With the form,
things are constrained already as is.

The extra:

as-table implemented which prompted the above change
test added for as-table