anti-forgery support added + manual testing
ring.middleware.anti-forgery support added. The token has to be used
directly, as the helper function only returns a string, not hiccup, so
it becomes escaped in when rendered. A new meta function is added.

A dev environment for manual testing is also added. This revealed that
the hidden form name had to be changed in order to be parsed by
wrap-keyword-params.

:posted? is added to the form in post-process-form. This is used to
determine if errors should be parsed at all or not. The default
behaviour is that we don't process errors if the form has not been
posted. Ie, it's OK to put in wrong values from the get go. We only
show errors once there is a user interaction.