shoelace added to manual testing
POC that web components works perfectly well with ez-form. The
field/render function has been changed to one arity, with defform
adding the default fields to the meta information for the
form. Instead of overriding the forms, :extra-fields has been added as
an option, which will be merged with the [:meta :fields]. See
dev/dev.clj how it's done in practice.