render lookup updated and table row layout added
Render lookup now checks to see if the first key in the lookup is a
field. If there is no value in the field with the lookup we get a nil,
which is what we want, as otherwise the lookup just shows up in the
hiccup and gets rendered.

as-table has table-opts changed so that a `:attributes` key is used for the
table attributes. `:table-row` is a new function that can be used to
switch out what the table uses for its rows.