Errors decoupled from fields
Errors have been moved to `[:meta :errors <k>]`. For validation `k` is
the field key and errors will show up there as normal. However, if you
wish to put in extra errors that do not have to do with any specific
field, it's now possible to do so by adding a new key in `[:meta
:errors]` and the refer to it as normal with `[<k> :errors :error]`.