:k added to field holding the :name value
:name is normalized when using qualified keywords, as the typical
Clojure webstack will not support qualified keywords when using a
standard <form> tag. This is converted back to the :name as it was
before, allowing for access to the fields according to the :name value
that was sent in.

If you need to go outside of ez-form, it is useful to be able to use
the unnormalized name.