#8 `:complex` validation added
Meant to be used for when validation and the error messages it
produces becomes too complex for ez-form to handle. Use cases includes
composite fields where a single field in turn is made up of many
fields.

The composite field is best thought of in terms of the value it should
return. It can be a map, or a sequence of maps, or a sequence of maps,
each containing one key that in turn holds a sequence of numbers.

Once this level of complexity is reached, it's better to hand it off
to the author of the field, while still providing the mechanisms for
validating the field.