#8 Test added for `:complex` validation for defform
Semi-realistic implementation of a field for path permissions. The
field is a composite of `path`, which has to be a string, and
`permissions`, which is a set of "write", "read" and "execute". The
value of the field is in turn a vector of maps which hold `path` and
`permissions`.

The implementation uses malli for validation, and returns an
mu/explain-data value if the value is not valid. This is then used to
determine which index and which composite key has the error.